requirements.txt
Pillow
numpy
scipy
matplotlib